+++ mbo/container/limited_map.h
@@ -190,8 +190,10 @@ class LimitedMap final
 
   // Custom
 
+  using LimitedBase::at_index;  // Index based access, throws `std::out_of_range` if not found.
   using LimitedBase::index_of;  // Return 0-based index of `key`, or `npos` if not found.
   using LimitedBase::npos;      // Return value for `index_of` if `key` is not found.
+
 
   // Read/write access