+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.2.14
 
 * Changed `tstring::find_first_of` and `tstring::find_last_of` solely with `string_view::find` to solve ASAN issue.
+* Added `LimitedSet::at_index` and `LimitedMap::at_index`.
 
 # 0.2.13